Split apart README-rendering code
