+++ find_datalad_repos/__main__.py
@@ -1,21 +1,14 @@
-from collections import defaultdict
 import logging
 from operator import attrgetter
-from pathlib import Path
-from typing import Dict, List, Mapping, Optional, Set
+from typing import Dict, List, Optional, Set
 import click
 from click_loglevel import LogLevel
 from pydantic import BaseModel, Field
 from .config import README_FOLDER, RECORD_FILE
+from .core import RepoRecord, mkreadmes
 from .github import GHDataladRepo, GHDataladSearcher, get_github_token
 from .osf import OSFDataladRepo, OSFDataladSearcher
-from .tables import TableRow, make_table_file
 from .util import Status, commit, runcmd
-
-
-class RepoRecord(BaseModel):
-    github: List[GHDataladRepo] = Field(default_factory=list)
-    osf: List[OSFDataladRepo] = Field(default_factory=list)
 
 
 class GHCollectionUpdater(BaseModel):
@@ -161,43 +154,7 @@ def main(log_level: int, mode: Optional[str]) -> None:
         with open(RECORD_FILE, "w") as fp:
             print(record.json(indent=4), file=fp)
 
-    Path(README_FOLDER).mkdir(parents=True, exist_ok=True)
-    repos_by_org: Mapping[str, List[GHDataladRepo]] = defaultdict(list)
-    for repo in record.github:
-        repos_by_org[repo.owner].append(repo)
-    main_rows: List[TableRow] = []
-    for owner, repos in repos_by_org.items():
-        if len(repos) > 1:
-            with Path(README_FOLDER, f"{owner}.md").open("w") as fp:
-                main_rows.append(
-                    make_table_file(
-                        fp,
-                        owner,
-                        list(repos),  # Copy to make mypy happy
-                        show_ours=False,
-                    )
-                )
-        else:
-            main_rows.extend(repos)
-    with open("README.md", "w") as fp:
-        print("# GitHub", file=fp)
-        make_table_file(fp, "", main_rows, show_ours=True)
-        print(file=fp)
-        print("# OSF", file=fp)
-        active: List[OSFDataladRepo] = []
-        gone: List[OSFDataladRepo] = []
-        for osfrepo in record.osf:
-            if osfrepo.gone:
-                gone.append(osfrepo)
-            else:
-                active.append(osfrepo)
-        for title, repolist in [("Active", active), ("Gone", gone)]:
-            print(f"## {title}", file=fp)
-            if repolist:
-                for i, osfrepo in enumerate(repolist, start=1):
-                    print(f"{i}. [{osfrepo.name}]({osfrepo.url})", file=fp)
-            else:
-                print("No repositories found!", file=fp)
+    mkreadmes(record)
 
     if mode != "readme":
         runcmd("git", "add", RECORD_FILE, "README.md", README_FOLDER)
